+++ BALSAMIC/assets/scripts/csv_to_pdf.py
@@ -2,8 +2,9 @@
 from pathlib import Path
 
 import click
-import pdfkit
 from pandas import DataFrame, read_csv
+
+from BALSAMIC.utils.pdf_report import get_table_html, html_to_pdf
 
 
 @click.command()
@@ -11,67 +12,29 @@ from pandas import DataFrame, read_csv
     "csv_path", nargs=1, required=True, type=click.Path(exists=True, resolve_path=True)
 )
 @click.argument("pdf_path", nargs=1, required=True, type=click.Path(resolve_path=True))
-def csv_to_pdf(csv_path: str, pdf_path: str) -> None:
+@click.option(
+    "--delimiter",
+    type=click.STRING,
+    default=",",
+    show_default=True,
+    help="CSV file delimiter.",
+)
+@click.option(
+    "-h",
+    "--header",
+    is_flag=True,
+    help="Include this option if the CSV file has a header row.",
+)
+def csv_to_pdf(csv_path: str, pdf_path: str, delimiter: str, header: bool) -> None:
     """Convert CSV file to a PDF."""
-    df: DataFrame = read_csv(csv_path)
+    df: DataFrame = read_csv(filepath_or_buffer=csv_path, delimiter=delimiter)
     html_table: str = df.to_html(
-        index=False, na_rep="NA", justify="center", escape=False
+        index=False, na_rep="NA", justify="center", escape=False, header=header
     )
-    html_page: str = get_table_html_page(
+    html_page: str = get_table_html(
         html_table=html_table, table_name=Path(csv_path).stem
     )
     html_to_pdf(html_string=html_page, pdf_path=pdf_path)
-
-
-def html_to_pdf(
-    html_string: str,
-    pdf_path: str,
-    orientation: str = "landscape",
-    margin_top: str = "1.5cm",
-    margin_bottom: str = "1cm",
-    margin_left: str = "1cm",
-    margin_right: str = "1cm",
-    zoom: int = 1,
-) -> None:
-    """Create a PDF file from the content of an HTML string."""
-    pdfkit.from_string(
-        input=html_string,
-        output_path=pdf_path,
-        options={
-            "page-size": "A4",
-            "encoding": "UTF-8",
-            "orientation": orientation,
-            "zoom": zoom,
-            "margin-top": margin_top,
-            "margin-bottom": margin_bottom,
-            "margin-left": margin_left,
-            "margin-right": margin_right,
-            "enable-local-file-access": None,
-        },
-    )
-
-
-def get_table_html_page(html_table: str, table_name: str) -> str:
-    """Return HTML-rendered content with the provided HTML table."""
-    return f"""
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <style>
-                h2 {{text-align: center; padding: 10px;}}
-                table {{margin: 0 auto; border: 1px solid black; border-collapse: collapse; text-align: center;}}
-                th {{font-size: 12pt; padding: 5px; background: #cccccc;}}
-                td {{font-size: 10pt; padding: 5px;}}
-                tr {{page-break-inside: avoid;}}
-                tr:nth-child(even) {{background: #eeeeee;}}
-            </style>
-        </head>
-        <body>
-            <h2>{table_name}</h2>
-            {html_table}
-        </body>
-        </html>
-    """
 
 
 if __name__ == "__main__":